[mac] More cleanup
